requirements.txt
pandas
requests
beautifulsoup4
selenium
webdriver-manager
XlsxWriter
pyisemail
dnspython
openpyxl
psutil
matplotlib
Pillow